# public/docs/admin/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SignInUserData": {
				"description": "Represents the data structure for user sign-in response.\nThis class is used to parse and store information about the signed-in user, including account type and tokens.",
				"properties": {
					"accountType": {
						"$ref": "#/components/schemas/AccountType",
						"description": "The type of the user account, which can be either 'Admin' or 'Customer'."
					},
					"accessToken": {
						"type": "string",
						"description": "Access token for the signed-in user."
					},
					"refreshToken": {
						"type": "string",
						"description": "Refresh token for the signed-in user."
					}
				},
				"required": [
					"accountType",
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"accountType": "User",
					"accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
					"refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
				}
			},
...
			"SignInUserRes": {
				"description": "Represents the response payload for user sign-in.\nExtends the base ApiResponse class with a generic parameter of SignInUserData type.\nThis class is used to handle the response after successfully signing in a user.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SignInUserData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordReq": {
				"description": "Represents the request payload for resetting a user's password.\nThis class is used when a user is resetting their password and provides a new password.",
				"properties": {
					"newPassword": {
						"type": "string",
						"description": "The new password for the user account.\nIt must not be empty and should have a minimum length of 8 characters."
					}
				},
				"required": [
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"newPassword": "NewPassw0rd"
				}
			},
...
			"SendPasswordResetTokenRes": {
				"description": "Represents the response payload for sending a password reset token.\nExtends the base ApiResponse class with a generic parameter of Success type.\nThis class is used to handle the response after sending a password reset token.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SendPasswordResetTokenReq": {
				"description": "Represents the request payload for sending a password reset token.\nThis class is used when a user requests a password reset and provides their email address.",
				"properties": {
					"emailAddress": {
						"type": "string",
						"description": "Email address of the user requesting a password reset.\nIt should be a valid email address format."
					}
				},
				"required": [
					"emailAddress"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"emailAddress": "john.doe@example.com"
				}
			},
...
			"CheckPasswordResetTokenRes": {
				"description": "Represents the response payload for checking the validity of a password reset token.\nExtends the base ApiResponse class with a generic parameter of CheckPasswordResetTokenData type.\nThis class is used to handle the response after checking the validity of a password reset token.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CheckPasswordResetTokenData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRes": {
				"description": "Represents the response payload for resetting a user's password.\nExtends the base ApiResponse class with a generic parameter of Success type.\nThis class is used to handle the response after successfully resetting a user's password.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SignInUserReq": {
				"description": "Represents the request payload for user sign-in.\nThis class is used when a user is trying to sign in and provides their email address and password.",
				"properties": {
					"emailAddress": {
						"type": "string",
						"description": "Email address of the user signing in.\nIt should be a valid email address format."
					},
					"password": {
						"type": "string",
						"description": "Password of the user signing in."
					}
				},
				"required": [
					"emailAddress",
					"password"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"emailAddress": "john.doe@example.com",
					"password": "Password123"
				}
			},
...
			"Gender": {
				"description": "Enum representing the gender of a user.\nIt can be either 'Male' or 'Female'.",
				"enum": [
					"Male",
					"Female"
				],
				"type": "string"
			},
...
			"GetUserRes": {
				"description": "Represents the response payload for retrieving detailed user data.\nThis class extends the ApiResponse class with a generic parameter of DetailedUserData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DetailedUserData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserRes": {
				"description": "Represents the response payload for updating a user.\nExtends the base ApiResponse class with a generic parameter of Success type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserReq": {
				"description": "Represents the request payload for creating or updating a user.\nThis class is used when information about a user, such as personal details and contact information, is provided.",
				"properties": {
					"roleId": {
						"type": "string",
						"description": "The MongoDB ObjectId of the user's role.\nIt must not be empty and should be a valid MongoDB ObjectId."
					},
					"firstName": {
						"type": "string",
						"description": "The first name of the user.\nIt must not be empty."
					},
					"middleName": {
						"type": "string",
						"description": "The middle name of the user (optional)."
					},
					"lastName": {
						"type": "string",
						"description": "The last name of the user.\nIt must not be empty."
					},
					"avatar": {
						"$ref": "#/components/schemas/Resource",
						"description": "Avatar resource for the user (optional).\nIt can be an instance of the Resource class."
					},
					"cover": {
						"$ref": "#/components/schemas/Resource",
						"description": "Cover resource for the user (optional).\nIt can be an instance of the Resource class."
					},
					"emailAddress": {
						"type": "string",
						"description": "The email address of the user.\nIt must not be empty and should be a valid email address format."
					},
					"phoneNumber": {
						"type": "string",
						"description": "The phone number of the user.\nIt must not be empty."
					},
					"dateOfBirth": {
						"type": "string",
						"description": "The date of birth of the user (optional).\nIt can be a string representing the date."
					},
					"gender": {
						"$ref": "#/components/schemas/Gender",
						"description": "The gender of the user (optional).\nIt should be one of the values from the Gender enum."
					},
					"citizenship": {
						"type": "string",
						"description": "The citizenship of the user (optional).\nIt should be a valid ISO 3166-1 alpha-3 country code."
					},
					"countryCode": {
						"type": "string",
						"description": "The ISO 3166-1 alpha-3 country code of the user.\nIt must not be empty and should be a valid ISO 3166-1 alpha-3 country code."
					},
					"stateId": {
						"type": "number",
						"format": "double",
						"description": "The ID of the state the user resides in (optional).\nIt should be a valid integer."
					},
					"cityId": {
						"type": "number",
						"format": "double",
						"description": "The ID of the city the user resides in (optional).\nIt should be a valid integer."
					},
					"addressLine": {
						"type": "string",
						"description": "The address line of the user (optional)."
					},
					"postalCode": {
						"type": "string",
						"description": "The postal code of the user (optional)."
					}
				},
				"required": [
					"roleId",
					"firstName",
					"lastName",
					"emailAddress",
					"phoneNumber",
					"countryCode"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"roleId": "5f4a7b61e55b8439b5bea1b2",
					"firstName": "John",
					"middleName": "Doe",
					"lastName": "Smith",
					"avatar": {
						"url": "https://example.com/avatar.jpg"
					},
					"cover": {
						"url": "https://example.com/cover.jpg"
					},
					"emailAddress": "john.doe@example.com",
					"phoneNumber": "1234567890",
					"dateOfBirth": "1990-01-01",
					"gender": "Male",
					"citizenship": "USA",
					"countryCode": "USA",
					"stateId": 1,
					"cityId": 123,
					"addressLine": "123 Main Street",
					"postalCode": "12345"
				}
			},
...
			"DeleteUserRes": {
				"description": "Represents the response payload for deleting a user.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Success"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"AccessToken": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			},
			"RefreshToken": {
				"type": "apiKey",
				"name": "x-refresh-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			}
		}
	},
	"info": {
		"title": "Node.js tsoa MongoDB Boilerplate Admin API",
		"version": "v1",
		"description": "This is a Node.js tsoa MongoDB Boilerplate Admin API version v1. It provides comprehensive documentation and administrative functionalities for managing MongoDB databases. The API is designed to streamline administrative tasks and enhance the overall efficiency of MongoDB operations. The 'public/docs/admin' directory serves as the output location for the API documentation, ensuring easy access for developers and administrators. With specVersion 3, this API adheres to the latest specifications, offering robust performance and compatibility. Whether it's managing user permissions, data schemas, or database configurations, this API simplifies administrative workflows and empowers developers to build scalable MongoDB applications with ease.",
		"license": {
			"name": "Apache 2.0"
		},
		"contact": {
			"name": "Akram B.",
			"email": "belhajsalem.akram@gmail.com"
		}
	},
	"paths": {
		"/auth/password/{token}": {
			"post": {
				"operationId": "setPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignInUserRes"
								}
							}
						}
					}
				},
				"description": "Set password for an admin.",
				"summary": "Set password for an admin using the provided token.",
				"tags": [
					"User Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The token associated with the admin for password reset.",
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing the new password.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordReq",
								"description": "- The request body containing the new password."
							}
						}
					}
				}
			}
		},
		"/auth/password-reset": {
			"post": {
				"operationId": "sendPasswordResetToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendPasswordResetTokenRes"
								}
							}
						}
					}
				},
				"description": "Send password reset token.",
				"summary": "Send a password reset token to the specified admin.",
				"tags": [
					"User Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing admin information for sending the reset token.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendPasswordResetTokenReq",
								"description": "- The request body containing admin information for sending the reset token."
							}
						}
					}
				}
			}
		},
		"/auth/password-reset/{token}": {
			"get": {
				"operationId": "checkPasswordResetToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckPasswordResetTokenRes"
								}
							}
						}
					}
				},
				"description": "Check password reset token validity.",
				"summary": "Check the validity of a password reset token.",
				"tags": [
					"User Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The password reset token to be checked.",
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "resetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordRes"
								}
							}
						}
					}
				},
				"description": "Reset admin password.",
				"summary": "Reset admin password using the provided token.",
				"tags": [
					"User Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The token associated with the admin for password reset.",
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing the new password.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordReq",
								"description": "- The request body containing the new password."
							}
						}
					}
				}
			}
		},
		"/auth/sign-in": {
			"post": {
				"operationId": "signInUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignInUserRes"
								}
							}
						}
					}
				},
				"description": "Sign in an admin.",
				"summary": "Sign in an admin.",
				"tags": [
					"User Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing admin credentials for sign-in.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignInUserReq",
								"description": "- The request body containing admin credentials for sign-in."
							}
						}
					}
				}
			}
		},
		"/me": {
			"get": {
				"operationId": "getCurrentAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserRes"
								}
							}
						}
					}
				},
				"description": "Get current admin.",
				"summary": "Get information about the currently authenticated admin.",
				"tags": [
					"Current User"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "updateCurrentAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateUserRes"
								}
							}
						}
					}
				},
				"description": "Update current admin.",
				"summary": "Update information for the currently authenticated admin.",
				"tags": [
					"Current User"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The request body containing updated admin information.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReq",
								"description": "- The request body containing updated admin information."
							}
						}
					}
				}
			}
		},
...
		"/users/{userId}": {
			"get": {
				"operationId": "getUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserRes"
								}
							}
						}
					}
				},
				"description": "Get user by ID.",
				"summary": "Get information about a specific user by ID.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The ID of the user to retrieve.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "updateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateUserRes"
								}
							}
						}
					}
				},
				"description": "Updates information about a specific user.",
				"summary": "Update user details.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The unique identifier of the user.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated information for the user.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReq",
								"description": "- The updated information for the user."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteUserRes"
								}
							}
						}
					}
				},
				"description": "Deletes a specific user.",
				"summary": "Delete a user.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The unique identifier of the user to be deleted.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/admin/v1"
		}
	],
...
}